DRY fixtures

+++ Learning Python/2017.09.28-TILs/demo/shapes/point.py
@@ -5,9 +5,9 @@ class Point(object):
 
     def __init__(self, x, y, z):
         """Initializes a new Point instance."""
-        self.__x = int(x)
-        self.__y = int(y)
-        self.__z = int(z)
+        self.__x = float(x)
+        self.__y = float(y)
+        self.__z = float(z)
 
     @property
     def x_coord(self):
